Moved mouse handling function in `Window` into separate module
